Fix where getVisibleRange().top could return -1
`th` is only assigned to here. It was always subtracting the height of
the horizontal scroll bar, even if no horizontal scroll bar was drawn.
Later, in scrollTo(), when there is no scroll bar `scrollTop` is set
to `th - viewportH`. That will be exactly negative the scroll bar
height. `scrollTop` is then used in calculating the getVisibleRange.top
value by floor(scrollTop / rowHeight) = -1.

Fix is to subtract the scrollbar height from `th` only if the scroll
bar is visible.

DEB-7911